Add Temp Fix
This commit adds a temporary fix to the weather app so it still adds all information. The API is not giving me the current time that I need, so the background will not update to the night wallpapers. Once it is finally updated, forcastDay.date will need to be updated to use .date instead of .hour[0].time in the api-controller.js module.